requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymysql==1.1.0
cryptography==41.0.7
qiskit==0.45.0
qiskit-ibm-runtime==0.15.1
pandas==2.1.3
numpy==1.25.2
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
sqlalchemy==2.0.23